+++ Source/Creation/Frame/Content/ID3FrameCreatorsFactory.swift
@@ -5,20 +5,12 @@
 //  2018 Fabrizio Duroni.
 //
 
-// swiftlint:disable function_body_length
 // swiftlint:disable line_length
 
 import Foundation
 
-class ID3iTunesFrameCreatorsFactory {
-    static func make() -> [ID3FrameCreator] {
-        return []
-    }
-}
-
 class ID3FrameCreatorsFactory {
     static func make() -> [ID3FrameCreator] {
-        let paddingAdder = PaddingAdderToEndOfContentUsingNullChar()
         let frameConfiguration = ID3FrameConfiguration()
         let frameFromStringUTF16ContentCreator = ID3FrameFromStringContentCreatorWithUTF16EncodingFactory.make()
         let frameFromStringISO88591ContentCreator = ID3FrameFromStringContentCreatorWithISO88591EncodingFactory.make()
@@ -28,19 +20,9 @@ class ID3FrameCreatorsFactory {
             ID3AlbumArtistFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3ArtistFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3TitleFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3RecordingYearFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3RecordingDayMonthFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3RecordingHourMinuteFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3RecordingDateTimeFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration, timestampCreator: ID3TimestampCreator()),
             ID3GenreFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3TrackPositionFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3AttachedPicturesFramesCreator(
-                attachedPictureFrameCreator: ID3AttachedPictureFrameCreator(
-                    id3FrameConfiguration: frameConfiguration,
-                    id3AttachedPictureFrameConfiguration: ID3AttachedPictureFrameConfiguration(),
-                    frameHeaderCreator: ID3FrameHeaderCreatorChain.make()
-                )
-            ),
+            ID3AttachedPicturesFramesCreatorFactory.make(),
             ID3ComposerFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3ConductorFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3ContentGroupingFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
@@ -50,25 +32,12 @@ class ID3FrameCreatorsFactory {
             ID3FileOwnerFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3LyricistFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3MixArtistFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesGroupingFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesMovementNameFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesMovementIndexFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesMovementCountFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesPodcastCategoryFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesPodcastDescriptionFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesPodcastIDFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3iTunesPodcastKeywordsFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3PublisherFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3SubtitleFrameCreator(frameCreator: frameFromStringUTF16ContentCreator, id3FrameConfiguration: frameConfiguration),
             ID3DiscPositionFrameCreator(frameCreator: frameFromStringISO88591ContentCreator, id3FrameConfiguration: frameConfiguration),
-            ID3FramesWithLocalizedContentCreator(
-                localizedFrameNames: frameNamesWithLocalizedContent,
-                localizedFrameCreator: ID3LocalizedFrameCreator(
-                    id3FrameConfiguration: frameConfiguration,
-                    frameHeaderCreator: ID3FrameHeaderCreatorChain.make(),
-                    paddingAdder: paddingAdder
-                )
-            )
+            ID3FramesWithLocalizedContentCreatorFactory.make()
         ]
+        + ID3RecordingTimesFrameCreatorsFactory.make()
+        + ID3iTunesFrameCreatorsFactory.make()
     }
 }
